🍻 Beer: update cart

+++ src/store/modules/cart.ts
@@ -2,13 +2,14 @@
  * @Author: CHENJIE
  * @Date: 2022-10-07 10:01:13
  * @LastEditors: CHENJIE
- * @LastEditTime: 2022-10-09 16:45:15
+ * @LastEditTime: 2022-10-09 17:31:23
  * @FilePath: /src/store/modules/cart.ts
  * @Description:cart
  */
 import { CartItem, CartItemRes } from '@/types/cart'
 import request from '@/utils/request'
 import { defineStore } from 'pinia'
+import useStore from '..'
 
 enum API {
   addCart = '/member/cart',
@@ -22,34 +23,52 @@ export default defineStore('cart', {
   state: () => ({
     list: [] as CartItem[], // 购物车数组
   }),
-
+  persist: true,
   // 方法
   actions: {
     /**
      * 添加购物车
      * @param data
      */
-    async addCart(data: { skuId: string; count: number }) {
-      await request.post(API.addCart, data)
-      this.getCartList()
+    async addCart(data: CartItem) {
+      if (this.isLogin) {
+        await request.post(API.addCart, data)
+        this.getCartList()
+      } else {
+        // 1. 判断购物车列表中是否有该商品数据
+        const goods = this.list.find((item) => item.skuId === data.skuId)
+        if (goods) {
+          // 2. 如果商品存在，直接把数量加起来
+          goods.count += data.count
+        } else {
+          // 3. 如果商品不存在，直接添加该商品
+          this.list.unshift(data)
+        }
+      }
     },
     /**
      * 获取购物车列表
      */
     async getCartList() {
-      const res = await request.get<CartItemRes>(API.getCartList)
-      this.list = res.data.result
+      if (this.isLogin) {
+        const res = await request.get<CartItemRes>(API.getCartList)
+        this.list = res.data.result
+      }
     },
     /**
      * 删除购物车商品
      * @param skuIds
      */
     async deleteCart(skuIds: string[]) {
-      await request.delete(API.deleteCart, {
-        data: { ids: skuIds },
-      })
-      // 重新获取最新购物车列表
-      this.getCartList()
+      if (this.isLogin) {
+        await request.delete(API.deleteCart, {
+          data: { ids: skuIds },
+        })
+        // 重新获取最新购物车列表
+        this.getCartList()
+      } else {
+        this.list = this.list.filter((item) => !skuIds.includes(item.skuId))
+      }
     },
     /**
      * 修改购物车是否选中/商品数量
@@ -60,21 +79,40 @@ export default defineStore('cart', {
       skuId: string,
       data: { selected?: boolean; count?: number }
     ) {
-      await request.put(API.updateCart + skuId, data)
-      // 重新获取最新购物车列表
-      this.getCartList()
+      if (this.isLogin) {
+        await request.put(API.updateCart + skuId, data)
+        // 重新获取最新购物车列表
+        this.getCartList()
+      } else {
+        const goods = this.effectiveList.find((item) => item.skuId === skuId)
+        // 这里要判断是否等于 undefined 因为返回结果是布尔值
+        if (data.selected !== undefined) {
+          goods!.selected = data.selected
+        }
+        if (data.count) {
+          goods!.count = data.count
+        }
+      }
     },
     // 修改全选或者全不选
     async updateCartAllSelected(isSelected: boolean) {
-      await request.put(API.updateCartAllSelected, {
-        selected: isSelected,
-      })
-      // 获取购物车列表
-      this.getCartList()
+      if (this.isLogin) {
+        await request.put(API.updateCartAllSelected, {
+          selected: isSelected,
+        })
+        // 获取购物车列表
+        this.getCartList()
+      } else {
+        this.effectiveList.forEach((item) => (item.selected = isSelected))
+      }
     },
   },
   // 计算
   getters: {
+    isLogin(): boolean {
+      const { user } = useStore()
+      return !!user.profile.token
+    },
     /**
      * 有效商品列表
      */
